### 2025-10-23 #9293 @editorialbot
[REVIEW]: OnlineNMF.jl: A Julia Package for Out-of-core and Sparse Non-negative Matrix Factorization
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/OnlineNMF.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.99.6
