+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM julia:1.8.0-rc1-buster
+FROM julia:1.9-buster
 
 ENV JULIA_DEPOT_PATH=/usr/local/julia
 
