### 2025-12-09 7337edf
Fix for JOSS review (docs and tests)
