+++ paper/paper.md
@@ -15,7 +15,7 @@ affiliations:
    index: 1
  - name: Laboratory for Bioinformatics Research, RIKEN Center for Biosystems Dynamics Research, Japan
    index: 2
-date: 8 July 2025
+date: 3 December 2025
 bibliography: paper.bib
 ---
 
@@ -27,7 +27,7 @@ Despite its broad applicability, NMF becomes computationally prohibitive for lar
 
 # Statement of need
 
-NMF is a workhorse algorithm for most data science tasks. However, as the size of the data matrix increases, it often becomes too large to fit into memory. In such cases, an out-of-core (OOC) implementation — where only subsets of data stored on disk are loaded into memory for computation — is desirable. Additionally, representing the data in a sparse matrix format, where only non-zero values and their coordinates are stored, is computationally advantageous. Therefore, a NMF implementation that supports both OOC computation and sparse data handling is highly desirable.
+NMF is a workhorse algorithm for most data science tasks. However, as the size of the data matrix increases, it often becomes too large to fit into memory. In such cases, an out-of-core (OOC) implementation — where only subsets of data stored on disk are loaded into memory for computation — is desirable. Additionally, representing the data in a sparse matrix format, where only non-zero values and their coordinates are stored, is computationally advantageous. Therefore, a NMF implementation that supports both OOC computation and sparse data handling is highly desirable (Figure 1).
 
 Similar discussions have been made in the context of Principal Component Analysis (PCA), and we have independently developed a Julia package, \texttt{OnlinePCA.jl} [@onlinepcajl]. \texttt{OnlineNMF.jl} is a spin-off version of \texttt{OnlinePCA.jl}, implementing NMF.
 
@@ -39,21 +39,19 @@ NMF can be easily reproduced on any machine where Julia is pre-installed by usin
 
 ## Installation
 
-First, install \texttt{OnlinePCA.jl} and \texttt{OnlineNMF.jl} from the official Julia package registry or directly from GitHub:
+First, install \texttt{OnlineNMF.jl} from the official Julia package registry or directly from GitHub:
 
 ```julia
-# Install OnlinePCA.jl and OnlineNMF.jl from Julia General
-julia> Pkg.add("OnlinePCA")
+# Install OnlineNMF.jl from Julia General
 julia> Pkg.add("OnlineNMF")
 
 # or GitHub for the latest version
-julia> Pkg.add(url="https://github.com/rikenbit/OnlinePCA.jl.git")
 julia> Pkg.add(url="https://github.com/rikenbit/OnlineNMF.jl.git")
 ```
 
 ## Preprocess of CSV
 
-Then, write a synthetic data as a CSV file, convert it to a compressed binary format using Zstandard, and prepare summary statistics for PCA. MM format is also supported for sparse matrices.
+Then, write a synthetic data as a CSV file, convert it to a compressed binary format using Zstandard, and prepare summary statistics for PCA. Matrix Market (MM) format is also supported for sparse matrices.
 
 ```julia
 using OnlinePCA
@@ -76,12 +74,10 @@ write_csv(joinpath(tmp, "Data.csv"), data)
 mmwrite(joinpath(tmp, "Data.mtx"), sparse(data))
 
[…]
