+++ paper/paper.bib
@@ -117,5 +117,6 @@
   title        = {{RcppPlanc}: R wrapper for the PLANC Nonnegative Matrix Factorization library},
   year         = {2023},
   howpublished = {\url{https://github.com/welch-lab/RcppPlanc}},
+  doi = {10.32614/CRAN.package.RcppPlanc},
   note         = {Accessed: 2025-05-01}
 }
