+++ src/dnmf.jl
@@ -306,13 +306,29 @@ function init_dnmf(
     lower::Number,
     upper::Number
 )
-    # Type Check
+    # Initizalization
     N, M = nm(input)
     binu = convert(Float32, binu)
     binv = convert(Float32, binv)
     teru = convert(Float32, teru)
     terv = convert(Float32, terv)
     graphv = convert(Float32, graphv)
+    # Check non-negative parameters
+    if binu < 0
+        throw(ArgumentError("binu must be non-negative, got $binu"))
+    end
+    if binv < 0
+        throw(ArgumentError("binv must be non-negative, got $binv"))
+    end
+    if teru < 0
+        throw(ArgumentError("teru must be non-negative, got $teru"))
+    end
+    if terv < 0
+        throw(ArgumentError("terv must be non-negative, got $terv"))
+    end
+    if graphv < 0
+        throw(ArgumentError("graphv must be non-negative, got $graphv"))
+    end
     # Initialization by NMF
     out_nmf = nmf(
         input=input,

[…]
