### 2025-12-09 9b62bb5
Fix for JOSS review (docs and tests)
