+++ .github/workflows/CI.yml
@@ -24,12 +24,13 @@ jobs:
       matrix:
         version:
           - '1.10'
+          - '1.11'
         os:
           - ubuntu-latest
         arch:
           - x64
     steps:
-      - uses: actions/checkout@v6
+      - uses: actions/checkout@v5
       - uses: julia-actions/setup-julia@v2
         with:
           version: ${{ matrix.version }}
