### 2025-12-09 68e3b65
Fix for JOSS review (docs and tests)
