### 2025-12-09 #9293 @kokitsuyuzaki
@rahulkhorana Sorry for my delay.
I have addressed the issues in v0.99.7.
https://github.com/rikenbit/OnlineNMF.jl/compare/v0.99.6...v0.99.7

> Tests

I added the tests listed in the following issue:
https://github.com/rikenbit/OnlineNMF.jl/pull/18

Some tests were intentionally omitted. For example, tests for all-zero input or partially negative input were not included because this tool is designed for very large-scale data, and checking all input values would incur significant computational cost, defeating the primary purpose of fast computation. Instead, I added documentation to README.md advising users to ensure their input data meets the requirements.
I also fixed the random seeds in `test/runtests.jl`.

> Potential Bug

https://github.com/rikenbit/OnlineNMF.jl/pull/19
I checked the notebook and found that `csv2bin` expects non-negative integer input. When floating-point values are provided, the function appears to run but produces meaningless values, which causes errors downstream. Adding runtime checks to validate whether each value is an integer would significantly degrade performance, contradicting the tool's primary goal of fast computation. Therefore, I added an explanation to `README.md` instead.
