### 2025-12-23 #9293 @crvernon
👋  @kokitsuyuzaki - we are almost there!

I just need you to address the following in your paper:

- [x] LINE 13:  "...and omics studies Rodriques (2019)." in this case should be formatted as "...and omics studies (Rodriques, 2019)."
- [x] LINE 17:  The wording in "...requirement, I originally developed OnlineNMF.jl, which is a Julia..." should be phrased as "...requirement,  I present OnlineNMF.jl, a Julia..."
- [x] LINE 39:  The header "Setting for plot" should be written as "Plot settings"
- [x] LINE 71:  The "p" in "python" should be capitalized.  You can do this in your bib file by wrapping any text you wish to maintain the formatting of in curly brackets.
- [x] LINE 88:  The "r" as in the R programming language should be capitalized.

- [x]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.

- [x]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).

- [x]  Please respond with the DOI of the archived version here along with the version tag name released.
