+++ paper/paper.bib
@@ -89,7 +89,7 @@
 @article{sklearn,
   author = {Pedregosa, F. et al.},
   journal = {Journal of Machine Learning Research},
-  title = {Scikit-learn: Machine learning in Python},
+  title = {Scikit-learn: Machine learning in {P}ython},
   volume = {12(85)},
   pages = {2825–2830},
   year = {2011}}
@@ -97,7 +97,7 @@
 @article{nntensor,
   author = {Tsuyuzaki, K. et al.},
   journal = {Journal of Open Source Software},
-  title = {nnTensor: An R package for non-negative matrix/tensor decomposition},
+  title = {nnTensor: An {R} package for non-negative matrix/tensor decomposition},
   volume = {8(84)},
   pages = {5015},
   year = {2023},
