### 2026-01-23 #9293 @crvernon
@kokitsuyuzaki there are still some issues with your Zenodo archive:

- [x] The title of the Zenodo archive should match you paper title exactly.
- [x] The license declared in the Zenodo archive should match the one you declare in your software repository.

Please edit the metadata in the Zenodo archive to make these changes.  A new release is not necessary.

Let me know when these changes have been made.  Thank you.
