### 2026-01-26 #9293 @kokitsuyuzaki
Thanks.
I've changed the title and license terms.
Please check them.
https://zenodo.org/records/18250632
