### 2026-01-26 #9293 @crvernon
@openjournals/dev

I am not sure what this error is pointing to.  Could you please advise?

> ⚠️ Couldn't acccept/publish paper. [An error happened](https://github.com/openjournals/joss-papers/actions/runs/21358993250?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ID figU003Anmf already defined
