### 2026-01-27 #9293 @arfon
This is because the paper.md file has two figures labeled "fig:nmf". The figure labels must be unique.
