### 2026-01-29 #9293 @kokitsuyuzaki
@crvernon Sorry for the inconvenience. I fixed the figure labels in the commit below.
https://github.com/rikenbit/OnlineNMF.jl/commit/4c94c02d33d0b7df83f332ed1af7fbea632cc3f1
