+++ README.md
@@ -3,6 +3,7 @@ Online Non-negative Matrix Factorization
 
 [![Build Status](https://github.com/rikenbit/OnlineNMF.jl/actions/workflows/CI.yml/badge.svg?branch=master)](https://github.com/rikenbit/OnlineNMF.jl/actions/workflows/CI.yml?query=branch%3Amaster)
 [![DOI](https://zenodo.org/badge/170862070.svg)](https://doi.org/10.5281/zenodo.8199413)
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.09293/status.svg)](https://doi.org/10.21105/joss.09293)
 
 ## 📚 Documentation
 [![Documentation](https://img.shields.io/badge/docs-dev-blue.svg)](https://rikenbit.github.io/OnlineNMF.jl/dev)
